Introducing a load spinner and some hook refactorings along the way
