+++ src/AppContextProvider.tsx
@@ -1,9 +1,7 @@
 import React, { useState, createContext, useEffect } from 'react'
 import { Company, CompaniesResource } from './types/Company'
-import { InteractionStatus } from '@azure/msal-browser'
-import { acquireToken } from './hooks/auth'
+import { useAcquireToken } from './hooks/auth'
 import { getHeaders, HttpClient } from './HttpClient'
-import { useIsAuthenticated, useMsal } from '@azure/msal-react'
 import { parseJwt } from './util/jwt'
 
 export type AppContextType = {
@@ -16,37 +14,9 @@ interface AppContextProviderProps {
 }
 
 const AppContextProvider = ({ children }: AppContextProviderProps) => {
-  const isAuthenticated = useIsAuthenticated()
-  const { instance, inProgress } = useMsal()
+  const [accessToken] = useAcquireToken()
   const [companies, setCompanies] = useState<Company[] | undefined>(undefined)
   const [roles, setRoles] = useState<string[] | undefined>(undefined)
-  const [accessToken, setAccessToken] = useState<string | undefined>(undefined)
-
-  /**
-   * If user is authenticated, let's fetch their company/azure ad info
-   */
-  useEffect(() => {
-    let ignore = false
-    if (inProgress === InteractionStatus.None && !ignore && !accessToken && isAuthenticated) {
-      acquireToken(instance, inProgress).then(
-        (tokenResult) => {
-          if (!tokenResult) {
-            // TODO: At some point, show some error notification
-            return
-          }
-          setAccessToken(tokenResult.accessToken)
-        },
-        (error) => {
-          // TODO: show alert
-          return Promise.reject(error)
-        }
-      )
-    }
-
-    return () => {
-      ignore = true
-    }
-  }, [instance, inProgress, isAuthenticated, accessToken])
 
   useEffect(() => {
     let ignore = false
