+++ src/components/DataDelivery/DataDeliveryView.tsx
@@ -1,6 +1,5 @@
 import './_dataDelivery.scss'
-import { useCallback, useEffect, useState } from 'react'
-import { FdsInputChange } from '../../../coreui-components/src/fds-input'
+import { useEffect, useRef, useState } from 'react'
 import { FdsInputComponent } from '../fds/FdsInputComponent'
 import { FdsButtonComponent } from '../fds/FdsButtonComponent'
 import { filterTableRowsBySearchWord, getTableHeaders, getTableRow } from './helpers'
@@ -9,6 +8,7 @@ import { useTranslation } from 'react-i18next'
 import { CompanyLatestEntryResource } from '../../types/DataDelivery'
 import Pagination from '../Common/Pagination/Pagination'
 import TableGroupedByColumn from '../Common/TableGroupedByColumn/TableGroupedByColumn'
+import { useSearchInputListener } from '../../hooks/searchInputListener'
 
 interface DataDeliveryProps {
   data: CompanyLatestEntryResource[] | null
@@ -16,7 +16,8 @@ interface DataDeliveryProps {
 
 const DataDeliveryView = ({ data }: DataDeliveryProps) => {
   const { t } = useTranslation()
-  const [searchWord, setSearchWord] = useState<string | null>(null)
+  const searchInputRef = useRef<HTMLDivElement | null>(null)
+  const [searchWord] = useSearchInputListener(searchInputRef, data)
   const [allRows, setAllRows] = useState<TableItem[][] | null>(null)
   const [rowsToShow, setRowsToShow] = useState<TableItem[][] | null>(null)
   const headerItems: HeaderItem[] = getTableHeaders(t)
@@ -47,27 +48,11 @@ const DataDeliveryView = ({ data }: DataDeliveryProps) => {
     }
   }, [data, t])
 
-  const useInputListener: EventListenerOrEventListenerObject = useCallback((e: Event) => {
-    const detail = (e as CustomEvent).detail as FdsInputChange
-    setSearchWord(detail.value as string)
-  }, [])
-
-  useEffect(() => {
-    const searchElement = document.querySelector('[id="searchInput"]')
-    if (searchElement && searchElement.getAttribute('listener') !== 'true') {
-      searchElement.addEventListener('change', useInputListener)
-    }
-
-    return () => {
-      searchElement?.removeEventListener('change', useInputListener)
-    }
-  }, [useInputListener])
-
   return (
     <>
       <div className={'searchLatestEntries'}>
         <form>
-          <div id={'searchInput'} className={'search-input'}>
+          <div id={'searchInput'} ref={searchInputRef} className={'search-input'}>
             <FdsInputComponent
               clearable={true}
               name={'searchWord'}
