+++ src/pages/AdminTools/DataDelivery/DataDeliveryPage.tsx
@@ -1,55 +1,27 @@
 import { useTranslation } from 'react-i18next'
-import { acquireToken, useAcquireToken } from '../../hooks/auth'
+import { acquireToken, useAcquireToken } from '../../../hooks/auth'
 import { AuthenticatedTemplate, UnauthenticatedTemplate, useMsal } from '@azure/msal-react'
-import { getHeaders, HttpClient } from '../../HttpClient'
-import { CompanyLatestEntryResource } from '../../types/DataDelivery'
-import { AppContext, AppContextType } from '../../AppContextProvider'
-import { rolesContainVacoAdmin, rolesContainVacoCompanyAdmin } from '../../util/role'
-import AuthRequiredPage from '../Error/AuthRequiredPage'
-import AdminRoleRequiredPage from '../Error/AdminRoleRequiredPage'
-import DataDeliveryView from '../../components/DataDelivery/DataDeliveryView'
-import { useContext, useEffect, useState } from 'react'
-import { FdsButtonVariant } from '../../../coreui-components/src/fds-button'
-import { FdsButtonComponent } from '../../components/fds/FdsButtonComponent'
-import { FdsTokenSize2 } from '../../../coreui-css/lib'
-import { downloadFile } from '../../util/download'
-import { getCurrentTimestamp } from '../../util/date'
+import { getHeaders, HttpClient } from '../../../HttpClient'
+import AuthRequiredPage from '../../Error/AuthRequiredPage'
+import AdminRoleRequiredPage from '../../Error/AdminRoleRequiredPage'
+import DataDeliveryView from '../../../components/DataDelivery/DataDeliveryView'
+import { FdsButtonVariant } from '../../../../coreui-components/src/fds-button'
+import { FdsButtonComponent } from '../../../components/fds/FdsButtonComponent'
+import { FdsTokenSize2 } from '../../../../coreui-css/lib'
+import { downloadFile } from '../../../util/download'
+import { getCurrentTimestamp } from '../../../util/date'
 import { AxiosResponse } from 'axios'
+import { useDataDeliveryFetch } from './hooks'
+import { useAdminRightsCheck } from '../hooks'
+import LoadSpinner from '../../../components/Common/LoadSpinner/LoadSpinner'
 
 const DataDeliveryPage = () => {
   const { t } = useTranslation()
   const { i18n } = useTranslation()
   const [accessToken] = useAcquireToken()
-  const appContext: AppContextType = useContext(AppContext)
+  const [data, isFetchInProgress] = useDataDeliveryFetch(accessToken)
+  const [hasAdminRole, hasCompanyAdminRole] = useAdminRightsCheck()
   const { instance, inProgress } = useMsal()
-  const [hasAdminRole, setHasAdminRole] = useState<boolean | undefined>(undefined)
-  const [hasCompanyAdminRole, setHasCompanyAdminRole] = useState<boolean | undefined>(undefined)
-  const [data, setData] = useState<CompanyLatestEntryResource[] | null>(null)
-
-  useEffect(() => {
-    if (appContext?.roles) {
-      setHasAdminRole(rolesContainVacoAdmin(appContext.roles))
-      setHasCompanyAdminRole(rolesContainVacoCompanyAdmin(appContext.roles))
-    }
-  }, [appContext])
-
-  useEffect(() => {
-    let ignore = false
-    if (accessToken && !ignore) {
-      HttpClient.get('/api/ui/admin/data-delivery', getHeaders(accessToken)).then(
-        (response) => {
-          setData(response.data as CompanyLatestEntryResource[])
-        },
-        (error) => {
-          // TODO: show alert
-          return Promise.reject(error)
-        }
-      )
-    }
-    return () => {
-      ignore = true
-    }
-  }, [accessToken])
 
   const handleDownloadedCsv = (response: AxiosResponse<any>) => {
     if (!response?.data) {
@@ -72,7 +44,6 @@ const DataDeliveryPage = () => {
               acquireToken(instance, inProgress).then(
                 (tokenResult) => {
                   if (!tokenResult) {
-                    // TODO: At some point, show some error notification
                     return
                   }
                   HttpClient.get(
@@ -90,7 +61,6 @@ const DataDeliveryPage = () => {
                   )
                 },
                 (error) => {
-                  // TODO: show alert
                   return Promise.reject(error)
                 }
               )
@@ -102,6 +72,7 @@ const DataDeliveryPage = () => {
             label={'Export CSV'}
           />
         </div>
+        {(hasAdminRole || hasCompanyAdminRole) && isFetchInProgress && <LoadSpinner />}
         {(hasAdminRole || hasCompanyAdminRole) && <DataDeliveryView data={data} />}
         {hasAdminRole !== undefined && hasCompanyAdminRole !== undefined && !(hasAdminRole || hasCompanyAdminRole) && (
           <AdminRoleRequiredPage />
