+++ src/pages/AdminTools/hooks.ts
@@ -1,0 +1,18 @@
+import { useContext, useEffect, useState } from 'react'
+import { rolesContainVacoAdmin, rolesContainVacoCompanyAdmin } from '../../util/role'
+import { AppContext, AppContextType } from '../../AppContextProvider'
+
+export const useAdminRightsCheck = () => {
+  const appContext: AppContextType = useContext(AppContext)
+  const [hasAdminRole, setHasAdminRole] = useState<boolean | undefined>(undefined)
+  const [hasCompanyAdminRole, setHasCompanyAdminRole] = useState<boolean | undefined>(undefined)
+
+  useEffect(() => {
+    if (appContext?.roles) {
+      setHasAdminRole(rolesContainVacoAdmin(appContext.roles))
+      setHasCompanyAdminRole(rolesContainVacoCompanyAdmin(appContext.roles))
+    }
+  }, [appContext])
+
+  return [hasAdminRole, hasCompanyAdminRole] as const
+}
